Add option to zoom to results
When `zoomToResults` is set to true, the map will zoom
upon completion of the query.
