+++ examples/desktop/app.js
@@ -133,7 +133,8 @@ app.loadMapbook({url: 'mapbook.xml'}).then(function() {
                 validateFieldValuesResult.message = 'Please complete at least one field.'
             }
             return validateFieldValuesResult;
-        }
+        },
+        zoomToResults: true
     });
