+++ src/services/search.js
@@ -75,6 +75,9 @@ function SearchService(Application, options) {
         };
     };
 
+    /** Automatically zoom to results */
+    this.zoomToResults = options.zoomToResults === true;
+
     /** Field transfomation function. */
     this.prepareFields = options.prepareFields ? options.prepareFields : function(fields) {
         // reformat the fields for the query engine,
@@ -145,7 +148,7 @@ function SearchService(Application, options) {
 
             // and footer contents.
             html += Application.renderTemplate(path, this.footerTemplate, query);
-            
+
         }
 
         // return the html for rendering.
