+++ src/services/select.js
@@ -75,6 +75,9 @@ function SelectService(Application, options) {
      */
     this.keepAlive = (options.keepAlive === true);
 
+    /** Automatically zoom to results */
+    this.zoomToResults = options.zoomToResults === true;
+
     /** User input fields, select allows choosing a layer */
     this.fields = options.fields || [{
         type: 'layers-list',
